requirements.txt
python-dotenv
langchain
langchain-google-genai
sentence-transformers
faiss-cpu
numpy
